mailservice/requirements.txt
Flask==3.0.2
Flask-Session==0.6.0
Flask-Cors==4.0.0
msal==1.28.0
requests==2.32.3
APScheduler==3.10.4
SQLAlchemy==2.0.25
python-dotenv==1.0.1
python-dateutil==2.8.2
gunicorn==21.2.0
